split up standard_kaldi so we can test
