+++ ext/Makefile
@@ -6,18 +6,19 @@ include $(KALDI_BASE)kaldi.mk
 
 BINFILES = standard_kaldi mkgraph
 
-OBJFILES =
+OBJFILES = decoder.o hypothesis.o hypothesizer.o rpc.o
 
 TESTFILES =
 
 CXXFLAGS += -I$(KALDI_BASE) -O3 -DNDEBUG
 
 ADDLIBS = $(KALDI_BASE)online2/kaldi-online2.a $(KALDI_BASE)ivector/kaldi-ivector.a \
-           $(KALDI_BASE)nnet2/kaldi-nnet2.a $(KALDI_BASE)lat/kaldi-lat.a \
+          $(KALDI_BASE)nnet2/kaldi-nnet2.a $(KALDI_BASE)lat/kaldi-lat.a \
           $(KALDI_BASE)decoder/kaldi-decoder.a  $(KALDI_BASE)cudamatrix/kaldi-cudamatrix.a \
           $(KALDI_BASE)feat/kaldi-feat.a $(KALDI_BASE)transform/kaldi-transform.a $(KALDI_BASE)gmm/kaldi-gmm.a \
           $(KALDI_BASE)thread/kaldi-thread.a $(KALDI_BASE)hmm/kaldi-hmm.a $(KALDI_BASE)tree/kaldi-tree.a \
           $(KALDI_BASE)matrix/kaldi-matrix.a $(KALDI_BASE)fstext/kaldi-fstext.a \
-          $(KALDI_BASE)util/kaldi-util.a $(KALDI_BASE)base/kaldi-base.a 
+          $(KALDI_BASE)util/kaldi-util.a $(KALDI_BASE)base/kaldi-base.a \
+          decoder.o hypothesis.o hypothesizer.o rpc.o
 
 include $(KALDI_BASE)makefiles/default_rules.mk
