+++ ext/standard_kaldi.cc
@@ -16,289 +16,12 @@
 #include "hmm/hmm-utils.h"
 #include "thread/kaldi-thread.h"
 
+#include "decoder.h"
+#include "hypothesis.h"
+#include "hypothesizer.h"
+#include "rpc.h"
+
 const int arate = 8000;
-
-// Phonemes contain information about phoneme timing in the
-// transcript. They're included in AlignedWords.
-struct Phoneme {
-  std::string token;
-  double duration;
-  bool has_duration;
-};
-
-// AlignedWord contains information about the timing of a
-// word heard in an audio file.
-struct AlignedWord {
-  std::string token;
-  double start;
-  bool has_start;
-  double duration;
-  bool has_duration;
-  std::vector<Phoneme> phones;
-};
-
-// A Hypothesis is a possible transcription of an auio file.
-// It's the output of this program.
-typedef std::vector<AlignedWord> Hypothesis;
-
-// Hypothesizer reads Kaldi lattices and extracts Hypothesis structs
-// with possible transcriptions of the utterances they contain.
-class Hypothesizer {
- public:
-  Hypothesizer(float frame_shift,
-               const kaldi::TransitionModel& transition_model,
-               const kaldi::WordBoundaryInfo& word_boundary_info,
-               const fst::SymbolTable* word_syms,
-               const fst::SymbolTable* phone_syms)
-      : frame_shift_(frame_shift),
-        transition_model_(&transition_model),
-        word_boundary_info_(&word_boundary_info),
-        word_syms_(word_syms),
-        phone_syms_(phone_syms) {}
-
-  // GetPartial only generates the word sequence for the best
-  // hypothesis in the lattice. It doesn't word-align the transcript
-  // or anything fancy. It's fast and good for partial results.
-  Hypothesis GetPartial(const kaldi::Lattice& lattice);
-
-  // GetFull extracts the best hypothesis in the lattice, pulling
-  // out all the stops. It word aligns, phoneme aligns, and generates
-  // confidences. It's slower and should only be used when it matters.
-  Hypothesis GetFull(const kaldi::Lattice& lattice);
-
- private:
-  float frame_shift_;
-  const kaldi::TransitionModel* transition_model_;
-  const kaldi::WordBoundaryInfo* word_boundary_info_;
-  const fst::SymbolTable* word_syms_;
-  const fst::SymbolTable* phone_syms_;
-};
-
-Hypothesis Hypothesizer::GetPartial(const kaldi::Lattice& lattice) {
-  Hypothesis hyp;
-
-  // Let's see what words are in here..
-  std::vector<int32> words;
-  std::vector<int32> alignment;
-  kaldi::LatticeWeight weight;
-  GetLinearSymbolSequence(lattice, &alignment, &words, &weight);
-
-  for (int32 word : words) {
-    AlignedWord aligned;
-    aligned.token = this->word_syms_->Find(word);
-    hyp.push_back(aligned);
-  }
-
-  return hyp;
-}
-
-Hypothesis Hypothesizer::GetFull(const kaldi::Lattice& lattice) {
-  Hypothesis hyp;
-
-  kaldi::CompactLattice clat;
-  ConvertLattice(lattice, &clat);
-
-  // Compute prons alignment (see: kaldi/latbin/nbest-to-prons.cc)
-  kaldi::CompactLattice aligned_clat;
-
-  std::vector<int32> words, times, lengths;
-  std::vector<std::vector<int32> > prons;
-  std::vector<std::vector<int32> > phone_lengths;
-
-  WordAlignLattice(clat, *this->transition_model_, *this->word_boundary_info_,
-                   0, &aligned_clat);
-
-  CompactLatticeToWordProns(*this->transition_model_, clat, &words, &times,
-                            &lengths, &prons, &phone_lengths);
-
-  for (int i = 0; i < words.size(); i++) {
-    AlignedWord aligned;
-    aligned.token = this->word_syms_->Find(words[i]);
-    aligned.start = times[i] * this->frame_shift_;
-    aligned.has_start = true;
-    aligned.duration = lengths[i] * this->frame_shift_;
-    aligned.has_duration = true;
-
-    for (size_t j = 0; j < phone_lengths[i].size(); j++) {
-      Phoneme phone;
-      phone.token = this->phone_syms_->Find(prons[i][j]);
-      phone.duration = phone_lengths[i][j] * this->frame_shift_;
-      phone.has_duration = true;
-      aligned.phones.push_back(phone);
-    }
-
-    hyp.push_back(aligned);
-  }
-
-  return hyp;
-}
-
-// Decoder represents an in-progress transcription of an utterance.
-class Decoder {
- public:
-  Decoder(const kaldi::OnlineNnet2FeaturePipelineInfo& info,
-          const kaldi::TransitionModel& transition_model,
-          const kaldi::OnlineNnet2DecodingConfig& nnet2_decoding_config,
-          const kaldi::nnet2::AmNnet& nnet,
-          const fst::Fst<fst::StdArc>* decode_fst,
-          const kaldi::OnlineIvectorExtractorAdaptationState& adaptation_state);
-
-  // AddChunk adds an audio chunk of audio to the decoding pipeline.
-  void AddChunk(kaldi::BaseFloat sampling_rate,
-                const kaldi::VectorBase<kaldi::BaseFloat>& waveform);
-  // GetAdaptationState gets the ivector extractor's adaptation state
-  void GetAdaptationState(
-      kaldi::OnlineIvectorExtractorAdaptationState* adaptation_state);
-  // GetBestPath outputs the decoder's current one-best lattice.
-  kaldi::Lattice GetBestPath();
-  // Finalize is called when you're finished adding chunks. It flushes the
-  // data pipeline and cleans up the lattice. After calling Finalize all
-  // calls to AddChunk will fail.
-  void Finalize();
-
- private:
-  // AdvanceDecoding processes any chunks in the decoding pipeline,
-  // applying silence weighting.
-  void AdvanceDecoding();
-
-  kaldi::OnlineNnet2FeaturePipeline feature_pipeline_;
-  kaldi::SingleUtteranceNnet2Decoder decoder_;
-
-  kaldi::OnlineSilenceWeighting silence_weighting_;
-  std::vector<std::pair<int32, kaldi::BaseFloat> > delta_weights_;
-
-  bool finalized_;
-};
-
-Decoder::Decoder(
-    const kaldi::OnlineNnet2FeaturePipelineInfo& info,
-    const kaldi::TransitionModel& transition_model,
-    const kaldi::OnlineNnet2DecodingConfig& nnet2_decoding_config,
-    const kaldi::nnet2::AmNnet& nnet,
-    const fst::Fst<fst::StdArc>* decode_fst,
-    const kaldi::OnlineIvectorExtractorAdaptationState& adaptation_state)
-    : feature_pipeline_(info),
-      decoder_(nnet2_decoding_config,
-               transition_model,
-               nnet,
-               *decode_fst,
-               &feature_pipeline_),
-      silence_weighting_(transition_model, info.silence_weighting_config),
-      finalized_(false) {
-  this->feature_pipeline_.SetAdaptationState(adaptation_state);
-}
-
-void Decoder::AddChunk(kaldi::BaseFloat sampling_rate,
-                       const kaldi::VectorBase<kaldi::BaseFloat>& waveform) {
-  if (finalized_) {
-    return;
-  }
-
-  this->feature_pipeline_.AcceptWaveform(sampling_rate, waveform);
-  this->AdvanceDecoding();
-}
-
-void Decoder::GetAdaptationState(
-    kaldi::OnlineIvectorExtractorAdaptationState* adaptation_state) {
-  this->feature_pipeline_.GetAdaptationState(adaptation_state);
-}
-
-void Decoder::AdvanceDecoding() {
-  // Down-weight silence in ivector estimation
-  if (this->silence_weighting_.Active()) {
-    this->silence_weighting_.ComputeCurrentTraceback(this->decoder_.Decoder());
-    this->silence_weighting_.GetDeltaWeights(
-        this->feature_pipeline_.NumFramesReady(), &this->delta_weights_);
-    this->feature_pipeline_.UpdateFrameWeights(this->delta_weights_);
-  }
-
-  this->decoder_.AdvanceDecoding();
-}
-
-kaldi::Lattice Decoder::GetBestPath() {
-  kaldi::Lattice lattice;
-
-  if (this->decoder_.NumFramesDecoded() == 0) {
-    return lattice;
-  }
-
-  this->decoder_.GetBestPath(this->finalized_, &lattice);
-
-  return lattice;
-}
-
-void Decoder::Finalize() {
-  this->finalized_ = true;
-  this->feature_pipeline_.InputFinished();
-  this->AdvanceDecoding();
-  this->decoder_.FinalizeDecoding();
-}
-
-// MarshalPhones serializes a list of phonemes as JSON
-std::string MarshalPhones(const vector<Phoneme>& phones) {
-  std::stringstream ss;
-
-  ss << "[";
-  for (int i = 0; i < phones.size(); i++) {
-    Phoneme phone = phones[i];
-    ss << "{" << std::endl;
-
-    ss << "\"phone\":\"" << phone.token << "\",";
-    if (phone.has_duration) {
-      ss << "\"duration\":" << phone.duration;
-    }
-
-    ss << "}";
-    if (i < phones.size() - 1) {
-      ss << ",";
-    }
-  }
-  ss << "]";
-
-  return ss.str();
-}
-
-// MarshalHypothesis serializes a Hypothesis struct as JSON
-std::string MarshalHypothesis(const Hypothesis& hypothesis) {
-  std::stringstream ss;
-
-  Hypothesis no_eps;
-  for (AlignedWord word : hypothesis) {
-    if (word.token == "<eps>") {
-      // Don't output anything for <eps> links, which correspond to silence....
-      continue;
-    }
-    no_eps.push_back(word);
-  }
-
-  ss << "{\"hypothesis\":";
-  ss << "[";
-
-  for (int i = 0; i < no_eps.size(); i++) {
-    AlignedWord word = no_eps[i];
-
-    ss << "{";
-
-    ss << "\"word\":\"" << word.token << "\",";
-    if (word.has_start) {
-      ss << "\"start\":" << word.start << ",";
-    }
-    if (word.has_duration) {
-      ss << "\"duration\":" << word.duration << ",";
-    }
-    ss << "\"phones\":" << MarshalPhones(word.phones);
-
-    ss << "}";
-    if (i < no_eps.size() - 1) {
-      ss << ",";
-    }
-  }
-
-  ss << "]";
-  ss << "}" << std::endl;
-
-  return ss.str();
-}
 
 void ConfigFeatureInfo(kaldi::OnlineNnet2FeaturePipelineInfo& info,
                        std::string ivector_model_dir) {
@@ -349,118 +72,6 @@ void ConfigEndpoint(kaldi::OnlineEndpointConfig& config) {
 
 void usage() {
   fprintf(stderr, "usage: standard_kaldi nnet_dir hclg_path proto_lang_dir\n");
-}
-
-// The status codes used by the RPC. Maps to HTTP status codes.
-enum {
-  STATUS_OK = 200,
-  STATUS_BAD_REQUEST = 400,
-  STATUS_INTERNAL_SERVER_ERROR = 500
-} RPCStatus;
-
-// Parse the method part of an RPC request. Returns false if the data is
-// malformed.
-//
-// Methods have this form:
-//   METHOD <ARG1> <ARG2> ... <ARGN>\n
-bool RPCReadMethod(std::istream& stream, string* method, vector<string>* args) {
-  std::string line;
-  if (!std::getline(stream, line)) {
-    return false;
-  }
-  std::stringstream ss(line);
-
-  if (!(ss >> *method)) {
-    return false;
-  }
-
-  string buf;
-  while (ss >> buf) {
-    args->push_back(buf);
-  }
-
-  return true;
-}
-
-class membuf : public std::basic_streambuf<char> {
- public:
-  membuf(std::vector<char>& vec) {
-    setg(vec.data(), vec.data(), vec.data() + vec.size());
-  }
-};
-
-// Read an RPC request. Returns false if the data is malformed.
-//
-// Methods have this form:
-//   MSG_SIZE\n
-//   METHOD <ARG1> <ARG2> ... <ARGN>\n
-//   BODY\n
-bool RPCReadRequest(std::istream& stream,
-                    string* method,
-                    vector<string>* args,
-                    vector<char>* body) {
-  string size_line;
-  if (!std::getline(stream, size_line)) {
-    return false;
-  }
-
-  std::stringstream size_ss(size_line);
-  size_t size;
-  if (!(size_ss >> size)) {
-    return false;
-  }
-
-  vector<char> request_bytes(size);
-  if (!stream.read(&request_bytes[0], size)) {
-    return false;
-  }
-  char trailing_newline;
-  if (!stream.read(&trailing_newline, 1)) {
-    return false;
-  }
-  if (trailing_newline != '\n') {
-    return false;
-  }
-
-  membuf request_buf(request_bytes);
-  std::istream request_stream(&request_buf);
-
-  if (!RPCReadMethod(request_stream, method, args)) {
-    return false;
-  }
-
-  // Read the rest into the body
-  body->insert(body->begin(), std::istreambuf_iterator<char>(request_stream),
-               {});
-
-  return true;
-}
-
-// Write the reply part of the RPC.
-//
-// Replies have this form:
-//   MSG_SIZE\n
-//   STATUS\n
-//   BODY\n
-void RPCWriteReply(std::ostream& stream,
-                   const int& status,
-                   const vector<char>& body) {
-  std::stringstream ss;
-  ss << status << std::endl;
-  ss.write(&body[0], body.size());
-  const string& reply_str = ss.str();
-
-  stream << reply_str.size() << std::endl;
-  stream.write(&reply_str[0], reply_str.size());
-  stream << std::endl;
-}
-
-// Write the reply part of the RPC. Same as the above but accepts strings.
-void RPCWriteReply(std::ostream& stream,
-                   const int& status,
-                   const string& body_str) {
-  vector<char> body(body_str.begin(), body_str.end());
-  RPCWriteReply(stream, status, body);
 }
 
 int main(int argc, char* argv[]) {
@@ -542,10 +153,15 @@ int main(int argc, char* argv[]) {
     vector<char> body;
 
     if (!RPCReadRequest(in_stream, &method, &args, &body)) {
-      RPCWriteReply(out_stream, STATUS_BAD_REQUEST,
-                    "malformed request '" + method + "'");
+      if (in_stream.eof()) {
+        RPCWriteReply(out_stream, STATUS_BAD_REQUEST, "unexpected eof");
+      } else {
+        RPCWriteReply(out_stream, STATUS_BAD_REQUEST,
+                      "malformed request '" + method + "'");
+      }
       continue;
     }
+    std::cerr << "method=" << method << std::endl;
 
     try {
       if (method == "stop") {
